Satellite Imagery Scene classification has been receiving a remarkable attention as it plays a vital role in a wide range of applications. However, the satellite imagery data size is quite large and often not all of the data is useful for many downstream tasks. We propose in this study to investigate the application of autoencoders to study the effectiveness of data compression on satellite imSatellite Imagery Scene classification has been receiving a remarkable attention as it plays a vital role in a wide range of applications. However, the satellite imagery data size is quite large and often not all of the data is useful for many downstream tasks. We propose in this study to investigate the application of autoencoders to study the effectiveness of data compression on satellite image data.age data.
Satellite Imagery Scene classification has been receiving a remarkable attention as it plays a vital role in a wide range of applications. However, the satellite imagery data size is quite large and often not all of the data is useful for many downstream tasks. We propose in this study to investigate the application of autoencoders and PCA to study the effectiveness of data compression on satellite image data.